pyproject.toml
[project]
name = "bdd-automation-ide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "behave>=1.3.3",
    "load-dotenv>=0.1.0",
    "ollama>=0.6.1",
    "playwright>=1.58.0",
    "pyside6>=6.10.2",
    "qtawesome>=1.4.1",
    "shiboken6>=6.10.2",
]
